add cost_filter

+++ stereo_matching/Cost.h
@@ -4,10 +4,10 @@
 #include "Solver.h"
 
 
-float SAD(Mat &ll, Mat &rr, Point l_pt, uchar disp, uchar win_h, uchar win_w, float* weight);
-float SSD(Mat &ll, Mat &rr, Point l_pt, uchar disp, uchar win_h, uchar win_w, float* weight);
+float SAD(Mat &ll, Mat &rr, Point l_pt, int disp, int win_h, int win_w, float* weight);
+float SSD(Mat &ll, Mat &rr, Point l_pt, int disp, int win_h, int win_w, float* weight);
 
-uint16_t CT(Mat &ll, Mat &rr, Point l_pt, uchar disp, uchar win_h, uchar win_w, float* weight);
-uint16_t hamming_cost(uint64_t ct_l, uint64_t ct_r);
+int CT(Mat &ll, Mat &rr, Point l_pt, int disp, int win_h, int win_w, float* weight);
+int hamming_cost(uint64_t ct_l, uint64_t ct_r);
 
-uint64_t CT_pts(Mat &im, int u, int v, uchar win_h, uchar win_w, float* weight);
+uint64_t CT_pts(Mat &im, int u, int v, int win_h, int win_w, float* weight);
