+++ stereo_matching/SGM.h
@@ -16,6 +16,6 @@ public:
 private:
 	float *L1, *L2, *L3, *L4, *L5, *L6, *L7, *L8;
 	float *min_L1, *min_L2, *min_L3, *min_L4, *min_L5, *min_L6, *min_L7, *min_L8;
-	uint16_t P1, P2;
+	int P1, P2;
 };
